Fixed passing query data to gocmdpev
If query data contains characters which can be interpreted by shell (ie. single quotes, or double quotes) than system call in gocmdpev method will be broken and it will return an error. Better way will be storing query data in temporary file and than cating it to gocmdpev (instead of echoing a string).